refactor(store-core): extract agent-tracking handlers into handlers/agent.ts (audit P2-3)
Eighth slice of the P2-3 split. Move the agent-tracking message handlers
(agent_spawned / agent_completed / agent_event) and the AgentInfoBuilder
interface into a new handlers/agent.ts.

Pure move, no logic change: the extracted block is byte-identical to the
removed lines (proven by diff). Imports AgentInfo + ChatMessage (../types,
type-only) and resolveSessionId (./_shared); AgentInfoBuilder is local to
the family. AgentInfo was used only by this family, so it is dropped from
the barrel's ../types import. handleBackgroundWorkChanged + its
PendingBackgroundShellsBuilder stay in index.ts (coupled to
buildSessionListPatches; they move with the session-list family later). The
barrel re-exports via `export * from './agent'`, so the public surface is
unchanged.

handlers/index.ts: 4868 -> 4696 lines.
